gateway/requirements.txt
# boto3==1.39.7
# python-dateutil>=2.8.2
# python-dotenv>=1.0.0

boto3>=1.24.0
python-dotenv>=1.0.1
requests>=2.31.0